mod: DepotElement and Investing
the keys of depot in Investing are now the hash of the user given name of the DepotElement.
This was done so that the name of one DepotElement could be used correctly in a id="" in an html tag, so that the JS Events know which DepotElement they are on.

The name for each DepotElement that the user choose, is now in this DepotElement.